Simplify page titles
Page Titles are a significant SEO factor.  We had included the
organization name "Hashrocket" into the page title to try to
differentiate our "Today I Learned" site from the others, but this
was an experiment rather than a solution.  With the tilex deployment
it was likely the wrong time to make that experiment.

In addition, according to the search console help:

https://support.google.com/webmasters/answer/35624?visit_id=1-636351220308427108-3910826740&rd=1

Branding is encouraged but boilerplate is discouraged.  It's unclear if
the Hashrocket part of "Today I Learned - Hashrocket" would be
interpreted as branding or as something different.

Conciseness is encouraged as well, therefore this commit reverts to a
page title style that is simpler and more like what we had previously in
the Rails version.