Increase cookie size to 8192 from default of 4096
Currently, likes are stored as cookies.  All the cookies are
collectively submitted as one http header.  Once you've made ~162 likes
then the header size for those cookies exceeds the max header size
allowed on the server.  This is also reportedly an issue in some
browsers but has not been encountered.  Chrome and Firefox handle large
headers fine.  Untested on IE.

For now up the header size allowed on the server to 8192.