Create a docker compose setup for tilex
This compose file serves two purposes. The first use case is to allow
building the tilex Docker image to later be pushed to DockerHub using
the `docker-compose build til` command.

The second use case is to run tilex locally without having to install
Erlang, Phoenix and PostgreSQL. This can be achieved with a
`docker-compose up` command and shortly after starting you'll find the
tilex app accessible on `locahost:4000`.