Make visitor_views_post test less brittle
The current test suite leave dirty data on the `requests` table when it
runs in parallel as there are some feature tests that ends up tracking
accesses outside of the Ecto Sandbox process for some tests.

So this commit intends to make this test less brittle by changing the
query to get the last created request on the db to compare the path.

I also changed the match to use the `=~` operator for better messaging
on errors.