Remove redundant MIME type
Per MDN:

Omitted or a JavaScript MIME type: This indicates the script is
JavaScript. The HTML5 specification urges authors to omit the attribute
rather than provide a redundant MIME type. In earlier browsers, this
identified the scripting language of the embedded or imported (via the
src attribute) code. JavaScript MIME types are listed in the
specification.

Source:

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/script