Forces oauther library version that deals with OTP 24
The newest versions of OTP do not have `:crypto.hmac/3` anymore and this
has broken TIL integration with Twitter since the last big app bump to
use OTP 24. `extwitter` library uses `oauther` which does not have
changed the `:crypto.mac/4` yet, so we're going to use this hot fix
branch for now.

For reference:
https://github.com/lexmag/oauther/pull/22
https://github.com/parroty/extwitter/issues/140